비밀번호 변경 free 기간을 90일로 늘림

+++ src/main/java/com/sbs/untactTeacher/util/Util.java
@@ -379,7 +379,7 @@ public class Util {
         }
     }
 
-    public static String getDateStrLater(int seconds) {
+    public static String getDateStrLater(long seconds) {
         SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
 
         String dateStr = format.format(System.currentTimeMillis() + seconds * 1000);
